consolidate transitions to single form
