+++ packages/esm-service-queues-app/translations/en.json
@@ -161,7 +161,6 @@
   "previousVisit": "Previous visit",
   "primaryHelperText": "Search for a patient",
   "priority": "Priority",
-  "priorityTransitioned": "Priority transitioned",
   "program": "Program",
   "provider": "Provider",
   "quantity": "Quantity",
@@ -170,16 +169,12 @@
   "queueEntryAddedSuccessfully": "Queue Entry Added Successfully",
   "queueEntryAddFailed": "Error adding queue entry status",
   "queueEntryError": "Error adding patient to the queue",
-  "queueEntryPriorityTransitionedSuccessfully": "Queue entry priority transitioned successfully",
-  "queueEntryPriorityTransitionFailed": "Error transitioning queue entry priority",
   "queueEntryRemoved": "Queue entry removed",
   "queueEntryRemovedSuccessfully": "",
-  "queueEntryStatusTransitionedSuccessfully": "Queue entry status transitioned successfully",
-  "queueEntryStatusTransitionFailed": "Error transitioning queue entry status",
   "queueEntryStatusUpdateFailed": "Error updating queue entry status",
-  "queueEntryTranferFailed": "Error transferring queue entry",
-  "queueEntryTransferred": "Queue entry transferred",
-  "queueEntryTransferredSuccessfully": "Queue entry transferred successfully",
+  "queueEntryTransitioned": "Queue entry transitioned",
+  "queueEntryTransitionedSuccessfully": "Queue entry transitioned successfully",
+  "queueEntryTransitionFailed": "Error transitioning queue entry",
   "queueEntryUpdateFailed": "Error updating queue entry status",
   "queueEntryUpdateSuccessfully": "Queue Entry Updated Successfully",
   "queueLocation": "Queue Location",
@@ -246,7 +241,6 @@
   "startVisitError": "Error starting visit",
   "startVisitQueueSuccessfully": "Patient has been added to active visits list and queue.",
   "status": "Status",
-  "statusTransitioned": "Status transitioned",
   "success": "Success",
   "tcaDate": "Tca date",
   "temperature": "Temperature",
@@ -255,9 +249,9 @@
   "tirageNotYetCompleted": "Triage has not yet been completed",
   "today": "Today",
   "transfer": "Transfer",
-  "transferToNextQueue": "Transfer to next queue",
-  "transitionPriority": "Transition priority",
-  "transitionStatus": "Transition status",
+  "transition": "Transition",
+  "transitionPatient": "Transition patient",
+  "transitionPatientStatusOrQueue": "Select a new status or queue for patient to transition to.",
   "triageForm": "Triage form",
   "triageNote": "Triage note",
   "trySearchWithPatientUniqueID": "Try searching with the patient's unique ID number",
